requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
openpyxl
Flask==3.0.0
line-bot-sdk==3.1.0